// angular/src/app/shared/models/configuration/InvSupervisor.ts
export class InvSupervisor {

      inv_supervisor_id: number;
      inv_supervisor_status: string;
      inv_supervisor_chk: number;
      inv_supervisor_name: string;
      inv_supervisor_shortname: string;
      inv_supervisor_type: string;
      inv_supervisor_legal_id: string;
      inv_supervisor_internal_id: string;
      inv_supervisor_schedule_id: string;
      inv_supervisor_schedule_name: string;

}